Resolve collection select issue when changing communities
Fixes an issue where the ajax call to refresh the collection options was running on every page change, instead of only when the community select changes. Further cleanup of this file may be warranted (the index code doesn't seem to serve a purpose, and if it does, should likely leverage a custom data attribute instead), but this doesn't impact the immediate bug. Closes #1012.